Added new TextFormater and child of it
TextFormater replaces old CenterText class that has been set to depricated and should be removed before merge into master. For the fighting system a child of TextFormater, FightTextFormater has been added, that implements a new print() that takes care of the display of health.

The TextFormater should be very versatile.