Updating Rooms (#16)
* Rooms (#7)

* Rooms added and defined

Added rooms, visit counter and first time events for certain rooms.

Created pseudocode for room items.

* Items and commandos

Added items to rooms and commandos to check inventory, use and take items

* Added REST command

Added the REST command

* Separated the source files from entities (#14)

* Separated the source files from entities

I have separated the source files from the newly added classes. This gives a more structured and clean development environment.

* Moved entities package outside the semestreprojekt package

I did this to match the proposed changes from #12

* Fight master (#6)

* Added fighting system basis

* Extended fighting system from prototype

Includes a tempMain, that needs to be deleted in merge with master.

Added two new classes for Characters and for attacks.

* Fight modifiers (#12)

* Added strength and accurary modifiers

Includes a couple of bugs that needs to be ironed out. Nameworthy is wrong attack damage.

Also renamed the Characters and Attacks classes to remove the pluralis

* Fixed attack damage not scaling

Attack damages now scale according to the strength of the enemy and the protagonist. Do mind that the values rounded.

Minor bug: Convertion to doubles still leaves the decimal value to 0

* Fixed missing precision

Attacks now correctly typecasts to double, truncating the precision to one decimal. Floored.

* Added accuracy for players and NPC

The hit/miss rate of players have been added based on the accuracy of the spell and the excess castTime.
For NPCs the rate is dependend on the accuracy of the spell and the reactiontime attribute.

Needs adjusting/balancing.

* Commented out debugging out prints

* Minor syntax updates for CenterText

* Added some javaDoc and victory/defeat message for the fights

* Added Flee option

The Flee option depends on the reactionstimes of both parts. You'll all ways flee from mobs with less reaction than you. You'll have an increasingly smaller chance at fleeing against mobs with higher reactiontime than you. (Mobs with 100 reactiontime more than you will be impossible to flee from)

* Added new TextFormater and child of it

TextFormater replaces old CenterText class that has been set to depricated and should be removed before merge into master. For the fighting system a child of TextFormater, FightTextFormater has been added, that implements a new print() that takes care of the display of health.

The TextFormater should be very versatile.

* Separated classes into their own packages

Development cleanup and organisation

* Added some javaDoc

* Flee chance inverted

I discovered that I inverted the flee chance based on the enemies reactionTime, so that lower reactionTimes (Values above 100) resulted in having a harder time fleeing.

* setBothPrint() added and comment formating

New method!

a combination of setBoth() and Print() allowing for quicker output with the TextFormater

* Minor javaDoc added